deleted the old generator file
